+++ src/engine_api.rs
@@ -1726,7 +1726,7 @@ impl Engine {
     ///
     /// let mut engine = Engine::new();
     ///
-    /// engine.on_progress(move |&ops| {
+    /// engine.on_progress(move |ops| {
     ///     if ops > 10000 {
     ///         Some("Over 10,000 operations!".into())
     ///     } else if ops % 800 == 0 {
@@ -1748,7 +1748,7 @@ impl Engine {
     #[inline(always)]
     pub fn on_progress(
         &mut self,
-        callback: impl Fn(&u64) -> Option<Dynamic> + SendSync + 'static,
+        callback: impl Fn(u64) -> Option<Dynamic> + SendSync + 'static,
     ) -> &mut Self {
         self.progress = Some(Box::new(callback));
         self