+++ src/fn_native.rs
@@ -343,18 +343,25 @@ pub type FnPlugin = dyn PluginFunction;
 #[cfg(feature = "sync")]
 pub type FnPlugin = dyn PluginFunction + Send + Sync;
 
-/// A standard callback function.
+/// A standard callback function for progress reporting.
 #[cfg(not(feature = "sync"))]
-pub type Callback<T, R> = Box<dyn Fn(&T) -> R + 'static>;
-/// A standard callback function.
+pub type OnProgressCallback = Box<dyn Fn(u64) -> Option<Dynamic> + 'static>;
+/// A standard callback function for progress reporting.
 #[cfg(feature = "sync")]
-pub type Callback<T, R> = Box<dyn Fn(&T) -> R + Send + Sync + 'static>;
+pub type OnProgressCallback = Box<dyn Fn(u64) -> Option<Dynamic> + Send + Sync + 'static>;
 
-/// A standard callback function.
+/// A standard callback function for printing.
+#[cfg(not(feature = "sync"))]
+pub type OnPrintCallback = Box<dyn Fn(&str) + 'static>;
+/// A standard callback function for printing.
+#[cfg(feature = "sync")]
+pub type OnPrintCallback<T, R> = Box<dyn Fn(&str) + Send + Sync + 'static>;
+
+/// A standard callback function for variable access.
 #[cfg(not(feature = "sync"))]
 pub type OnVarCallback =
     Box<dyn Fn(&str, usize, &EvalContext) -> Result<Option<Dynamic>, Box<EvalAltResult>> + 'static>;
-/// A standard callback function.
+/// A standard callback function for variable access.
 #[cfg(feature = "sync")]
 pub type OnVarCallback = Box<
     dyn Fn(&str, usize, &EvalContext) -> Result<Option<Dynamic>, Box<EvalAltResult>>